feat  : 프로필사진 표출

+++ src/main/java/com/runHani/repository/UserProfileFileRepository.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.runHani.entity.UserProfileFileEntity;
 
 
-public interface UserProfileFileRepository extends JpaRepository<UserProfileFileEntity, Long>  {
+public interface UserProfileFileRepository extends JpaRepository<UserProfileFileEntity, Integer>  {
